move bytecode circuit to conform with the codebase style (#1632)
### Description

We move zkevm-circuits/src/bytecode_circuit/circuit.rs to
zkevm-circuits/src/bytecode_circuit.rs

### Issue Link

#1607

### Type of change

Bug fix (non-breaking change which fixes an issue)


Before

```
wc -l zkevm-circuits/src/*circuit.rs
      12 zkevm-circuits/src/bytecode_circuit.rs
     887 zkevm-circuits/src/copy_circuit.rs
     528 zkevm-circuits/src/evm_circuit.rs
     550 zkevm-circuits/src/exp_circuit.rs
    1069 zkevm-circuits/src/keccak_circuit.rs
     796 zkevm-circuits/src/mpt_circuit.rs
    1653 zkevm-circuits/src/pi_circuit.rs
     190 zkevm-circuits/src/root_circuit.rs
     614 zkevm-circuits/src/state_circuit.rs
     493 zkevm-circuits/src/super_circuit.rs
     361 zkevm-circuits/src/tx_circuit.rs
    7153 total
```


After

```
wc -l zkevm-circuits/src/*circuit.rs
     828 zkevm-circuits/src/bytecode_circuit.rs
     887 zkevm-circuits/src/copy_circuit.rs
     528 zkevm-circuits/src/evm_circuit.rs
     550 zkevm-circuits/src/exp_circuit.rs
    1069 zkevm-circuits/src/keccak_circuit.rs
     796 zkevm-circuits/src/mpt_circuit.rs
    1653 zkevm-circuits/src/pi_circuit.rs
     190 zkevm-circuits/src/root_circuit.rs
     614 zkevm-circuits/src/state_circuit.rs
     491 zkevm-circuits/src/super_circuit.rs
     361 zkevm-circuits/src/tx_circuit.rs
    7967 total
```
