+++ zkevm-circuits/src/bytecode_circuit/test.rs
@@ -1,13 +1,9 @@
-use crate::{
-    bytecode_circuit::circuit::BytecodeCircuit,
-    util::{log2_ceil, unusable_rows, SubCircuit},
-};
+use super::{BytecodeCircuit, BytecodeCircuitRow};
+use crate::util::{log2_ceil, unusable_rows, SubCircuit};
 use bus_mapping::{evm::OpcodeId, state_db::CodeDB};
 use eth_types::Field;
 use halo2_proofs::{arithmetic::Field as Halo2Field, dev::MockProver, halo2curves::bn256::Fr};
 use log::error;
-
-use super::circuit::BytecodeCircuitRow;
 
 #[test]
 fn bytecode_circuit_unusable_rows() {
